Bugfix: Also take ratio into account when increasing step-size (Kotlin)
Not taking this into account could have risked missing a record-setter for
the ratio. Step-size increases more slowly now, so the records can only be
computed up to 1e11 in 25 minutes. However, they're still matching at least
that far.

(It's likely that larger steps can be made for the ratio after further
analysis.)