Use Weber 2020 for a faster (and proven) Zaremba function
The previous version here was a guess at repairing a typo in the Erdős and
Zaremba paper, but I got it wrong. (It matched the results, but we don't
have a proof for it!) The new version here draws from Weber and is faster
because h() can be lifted out of the inner loop.